[SNOW-937189] Split tests for old and new connection pool
